Extract SwapInProgress body as TxWidgetWithInfoMsg widget
